+++ rest/model/dns/zone.go
@@ -14,7 +14,10 @@ type Zone struct {
 	Pool         string   `json:"pool,omitempty"` // Deprecated
 
 	ID   string `json:"id,omitempty"`
-	Zone string `json:"zone,omitempty"`
+	Zone string `json:"zone,omitempty"` // FQDN
+	// Unique identifier for the zone within account. In most cases, it is best
+	// to also use the zone FQDN for this.
+	Name string `json:"name,omitempty"`
 
 	TTL        int    `json:"ttl,omitempty"`
 	NxTTL      int    `json:"nx_ttl,omitempty"`
@@ -41,10 +44,13 @@ type Zone struct {
 	// Whether or not DNSSEC is enabled on the zone. Note we use a pointer so
 	// leaving this unset will not change a previous setting.
 	DNSSEC *bool `json:"dnssec,omitempty"`
+
+	// Array of view names
+	Views []string `json:"views,omitempty"`
 }
 
 func (z Zone) String() string {
-	return z.Zone
+	return z.GetName()
 }
 
 // ZoneRecord wraps Zone's "records" attribute
@@ -110,9 +116,27 @@ type TSIG struct {
 // NewZone takes a zone domain name and creates a new zone.
 func NewZone(zone string) *Zone {
 	z := Zone{
+		Name: zone,
 		Zone: zone,
 	}
 	return &z
+}
+
+// NewNamedZone takes a zone name and FQDN and creates a new zone.
+func NewNamedZone(name string, zone string) *Zone {
+	z := Zone{
+		Zone: zone,
+		Name: name,
+	}
+	return &z
+}
+
+// GetName returns the Name, if present, other FQDN is the name
+func (z *Zone) GetName() string {
+	if z.Name != "" {
+		return z.Name
+	}
+	return z.Zone
 }
 
 // MakePrimary enables Primary, disables Secondary, and sets primary's